prediction_api/requirements.txt
aiohttp==2.2.5
async-timeout==1.4.0
bleach==1.5.0
chardet==3.0.4
html5lib==0.9999999
Keras==2.0.8
Markdown==2.6.9
multidict==3.2.0
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.4.0
h5py==2.7.1
PyYAML==3.12
scipy==0.19.1
six==1.11.0
Werkzeug==0.12.2
yarl==0.12.0
